% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.benthic.data.R
\name{cGENIE.benthic.data}
\alias{cGENIE.benthic.data}
\title{Summarize Bottom-Water Environments from cGENIE NetCDF Files}
\usage{
cGENIE.benthic.data(var, experiment, output = "dataframe")
}
\arguments{
\item{var}{Character string specifying the variable to extract from the NetCDF file.}

\item{experiment}{Character string indicating the path to the experiment directory containing the NetCDF files.}

\item{output}{Character. The format of the output. Default is "dataframe". Options are "dataframe" or "array".}
}
\value{
A data frame or 3D array with the original coordinates and matched climate data from the cGENIE model.
}
\description{
This function extracts and summarizes the bottom-water redox state from 3D cGENIE NetCDF files. It retrieves specified variables, computes the corresponding bottom-water depths, and optionally returns the data as either an array or a dataframe.
}
