% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.bw.redox.R
\name{cGENIE.bw.redox}
\alias{cGENIE.bw.redox}
\title{Summarize Bottom-Water Redox Conditions}
\usage{
cGENIE.bw.redox(
  experiment,
  anox.thresh = 0,
  subox.thresh = 4.8e-06,
  lower.anox.lim = TRUE
)
}
\arguments{
\item{experiment}{A character string specifying the path to the experiment folder, where the necessary NetCDF files are stored.}

\item{anox.thresh}{Numeric. The oxygen threshold (in mol/kg) for classifying anoxic bottom waters. Default is 0.}

\item{subox.thresh}{Numeric. The oxygen threshold (in mol/kg) for classifying suboxic bottom waters. Default is 4.8e-6 (based on Sperling et al. 2015).}
}
\value{
A list with the following components:
\describe{
\item{f.anox}{Fraction of bottom-water grid cells classified as anoxic.}
\item{f.subox}{Fraction of bottom-water grid cells classified as suboxic.}
\item{f.ox}{Fraction of bottom-water grid cells classified as oxic.}
\item{bw.array}{A matrix of bottom-water oxygen concentrations for each grid cell.}
}
}
\description{
This function processes output from a cGENIE experiment to summarize the redox conditions
of bottom-water environments. It categorizes the ocean bottom water into anoxic, suboxic, and oxic conditions
based on specified oxygen thresholds. The function also handles the topographic grid data to ensure correct depth assignments.
}
\details{
This function reads the NetCDF output of a cGENIE experiment, processes the ocean bottom-water oxygen data (\code{ocn_O2}), and
categorizes each bottom-water grid cell as anoxic, suboxic, or oxic. The topographic data (\code{grid_topo}) is used to align the depth of each grid cell with the
correct ocean layer, and the function outputs the fraction of anoxic, suboxic, and oxic bottom waters.

The redox classification thresholds are based on oxygen concentration values, with:
\itemize{
\item Anoxic: oxygen concentration \eqn{\leq} anox.thresh
\item Suboxic: oxygen concentration between anox.thresh and subox.thresh
\item Oxic: oxygen concentration \eqn{>} subox.thresh
}
}
\examples{
\dontrun{
# Example usage:
result <- cGENIE.bw.redox("path_to_experiment")
print(result$f.anox)
print(result$f.subox)
print(result$f.ox)
}

}
\references{
Sperling, E. A., et al. 2015 The Ecological Physiology of Earth's Second Oxygen Revolution.
}
