% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.data.R
\name{cGENIE.data}
\alias{cGENIE.data}
\title{Extract 2D Data Fields from 3D or 2D netCDF Files}
\usage{
cGENIE.data(
  var,
  experiment,
  depth.level = NULL,
  dims = 3,
  year = "default",
  model = "biogem"
)
}
\arguments{
\item{var}{The variable name to extract from the .nc file.}

\item{experiment}{Directory containing the experiment's netCDF files.}

\item{depth.level}{Depth level to extract the data from. Set to NULL by default.}

\item{dims}{Number of dimensions in the netCDF file. Defaults to 3.}

\item{year}{Year to extract data for (default is "default", meaning the last time point).}

\item{model}{The model type; defaults to "biogem".}
}
\value{
A data frame containing the extracted 2D lat-lon data field with corresponding grid information.
}
\description{
This function extracts 2D latitude-longitude data fields from 3D or 2D netCDF (.nc) files.
}
