% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.get.nc.R
\name{cGENIE.get.nc}
\alias{cGENIE.get.nc}
\title{Import Data from cGENIE NetCDF Files}
\usage{
cGENIE.get.nc(var, experiment, dims = 3, model = "biogem")
}
\arguments{
\item{var}{A character string specifying the name of the variable to extract from the NetCDF file.}

\item{experiment}{A character string indicating the path to the cGENIE experiment directory.}

\item{dims}{An integer specifying the dimensions of the NetCDF file to read. Default is set to 3 for 3D data but can alternatively be set to 2. .}

\item{model}{A character string indicating the model type (default is "biogem"). Currently, only "biogem" is supported.}
}
\value{
A list containing latitude, longitude, depth, time, and the specified variable data from the NetCDF file.
}
\description{
This function imports specified variables from cGENIE NetCDF files and extracts relevant grid information.
}
\details{
The function reads a NetCDF file based on the specified experiment and model, extracting latitude, longitude,
depth, time, and the requested variable. It supports both 2D and 3D datasets.
}
