% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.map.ortho.vid.R
\name{cGENIE.map.ortho.vid}
\alias{cGENIE.map.ortho.vid}
\title{Create an Orthographic Projection Video of cGENIE Map}
\usage{
cGENIE.map.ortho.vid(
  experiment,
  var,
  depth.level = 2,
  dims = 3,
  year = "default",
  unit.factor = NULL,
  min.value = NULL,
  max.value = NULL,
  intervals = NULL,
  continents.outlined = TRUE,
  line.thickness = 0.5,
  line.colour = "white",
  scale.label = NULL,
  model = "biogem",
  palette_name = pals::parula(1000),
  darkmode = TRUE,
  background.color = "black",
  text.color = "white",
  framerate = 10,
  lon_steps = 1,
  output_file = "output.mp4"
)
}
\arguments{
\item{experiment}{Character. The name of the experiment.}

\item{var}{Character. The variable to be plotted.}

\item{depth.level}{Numeric. The depth level to be used. Default is 2.}

\item{dims}{Numeric. The dimensions to be used. Default is 3.}

\item{year}{Character or Numeric. The year to be used. Default is "default".}

\item{unit.factor}{Numeric. A factor to convert units. Default is NULL.}

\item{min.value}{Numeric. The minimum value for the color scale. Default is NULL.}

\item{max.value}{Numeric. The maximum value for the color scale. Default is NULL.}

\item{intervals}{Numeric. The intervals for the color scale. Default is NULL.}

\item{continents.outlined}{Logical. Whether to outline continents. Default is TRUE.}

\item{line.thickness}{Numeric. The thickness of the outline lines. Default is 0.5.}

\item{line.colour}{Character. The color of the outline lines. Default is "grey20".}

\item{scale.label}{Character. The label for the color scale. Default is NULL.}

\item{model}{Character. The model to be used. Default is "biogem".}

\item{palette_name}{Character. The name of the color palette to be used. Default is \code{pals::parula(1000)}.}

\item{darkmode}{Logical. Whether to use dark mode. Default is FALSE.}

\item{background.color}{Character. The background color. Default is "black".}

\item{text.color}{Character. The text color. Default is "white".}

\item{framerate}{Numeric. The frame rate of the video. Default is 10.}

\item{lon_steps}{Numeric. The steps for longitude rotation. Default is 1.}

\item{output_file}{Character. The name of the output video file. Default is "output.mp4".}
}
\value{
None. The function creates a video file as specified by \code{output_file}.
}
\description{
This function generates a video of orthographic projections of cGENIE map data by rotating the longitude.
}
