% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.map.R
\name{cGENIE.map}
\alias{cGENIE.map}
\title{Generate Maps from cGENIE Model Output}
\usage{
cGENIE.map(
  var,
  experiment,
  depth.level = 1,
  dims = NULL,
  year = "default",
  unit.factor = NULL,
  min.value = NULL,
  max.value = NULL,
  intervals = NULL,
  continents.outlined = TRUE,
  scale.label = NULL,
  model = "biogem",
  line.thickness = 1,
  palette_name = pals::parula(1000),
  projection = "ESRI:54012",
  darkmode = FALSE,
  custom.bg = NULL,
  custom.fg = NULL,
  col.labels = NULL
)
}
\arguments{
\item{var}{(character) Variable to visualize (e.g. "ocn_temp", "ocn_sal", "ocn_O2").}

\item{experiment}{(character) Path or name of the experiment.}

\item{depth.level}{(numeric) Depth layer to visualize (default = 1).}

\item{dims}{(numeric) Dimensionality of the data (default NULL; auto-set by var).}

\item{year}{(numeric or character) Time step to visualize (default "default").}

\item{unit.factor}{(numeric) Scaling factor for variable (default NULL; auto-set by var).}

\item{min.value, max.value}{(numeric) Color scale limits (default NULL; auto-set by var).}

\item{intervals}{(numeric) Color scale intervals (default NULL; auto-set by var).}

\item{continents.outlined}{(logical) Whether to outline continents (default TRUE).}

\item{scale.label}{(character) Label for color scale (default NULL; auto-set by var).}

\item{model}{(character) Model type (default "biogem").}

\item{line.thickness}{(numeric) Thickness of continent outlines (default 1).}

\item{palette_name}{(character) Color palette for the plot (default pals::parula(1000)).}

\item{projection}{(character) Map projection (default 'ESRI:54012').}

\item{darkmode}{(logical) Enable dark mode (default FALSE).}

\item{custom.bg}{(character) Background color for dark mode (default NULL; auto-set if darkmode = TRUE).}

\item{custom.fg}{(character) Foreground color for dark mode (default NULL; auto-set if darkmode = TRUE).}

\item{col.labels}{(numeric) Labels for the color bar ticks.}
}
\value{
A ggplot object representing the map.
}
\description{
Generate Maps from cGENIE Model Output
}
