% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.MI.array.local2.R
\name{cGENIE.MI.array.local2}
\alias{cGENIE.MI.array.local2}
\title{cGENIE.MI.array.local2}
\usage{
cGENIE.MI.array.local2(
  experiment,
  year = "default",
  n_ecotypes = 1000,
  seed_number = 1993,
  format = "habitat_viability",
  env = NULL,
  shelf.depth = ifelse(env == "shelf", 3, NULL),
  surf = ifelse(env == "shelf", FALSE, NULL),
  time.step = ifelse(env == "shelf", "default", NULL),
  depth.level = ifelse(env == "points", NULL, NULL),
  coord.dat = ifelse(env == "points", coord.dat, NULL),
  lat.name = ifelse(env == "points", "p_lat", NULL),
  lng.name = ifelse(env == "points", "p_lng", NULL)
)
}
\arguments{
\item{experiment}{Character. The path to the experiment directory containing cGENIE output files.}

\item{year}{Character or numeric. The year to process. If "default", the final time step is used.}

\item{n_ecotypes}{Integer. The number of ecophysiotypes (ecotypes) to model. Default is 1000.}

\item{seed_number}{Integer. The seed number for random number generation. Default is 1993.}

\item{format}{Character. The format of the output. Either "habitat_viability" or "ecotype_viability". Default is "habitat_viability".}

\item{env}{Character. The environment type. Either "shelf" or "points". Default is NULL.}

\item{shelf.depth}{Numeric. The depth for shelf environments. Default is 3 if env is "shelf", otherwise NULL.}

\item{surf}{Logical. Whether to use surface data for shelf environments. Default is FALSE if env is "shelf", otherwise NULL.}

\item{time.step}{Character or numeric. The time step to use for shelf environments. Default is "default" if env is "shelf", otherwise NULL.}

\item{depth.level}{Numeric. The depth level for point environments. Default is NULL.}

\item{coord.dat}{Data frame. The coordinate data for point environments. Default is NULL.}

\item{lat.name}{Character. The name of the latitude column in coord.dat for point environments. Default is "p_lat".}

\item{lng.name}{Character. The name of the longitude column in coord.dat for point environments. Default is "p_lng".}
}
\value{
Depending on the format parameter, returns either a 3D array of habitat viability percentages or a 4D array of ecotype viability.
}
\description{
This function processes cGENIE model output to calculate habitat viability and ecotype viability based on metabolic index models.
}
\details{
This function reads cGENIE model output and calculates habitat viability and ecotype viability using metabolic index models. It sources functions from Hofmann et al. (2011) and uses parameter values from Penn et al. (2018). The function processes the data to generate arrays of partial pressure of oxygen (pO2) and temperature in Kelvin, and then calculates the viability of each ecotype in each model cell.
}
