% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.MI.array.R
\name{cGENIE.MI.array}
\alias{cGENIE.MI.array}
\title{cGENIE.MI.array.R}
\usage{
cGENIE.MI.array(
  experiment,
  year = "default",
  n_ecotypes = 1000,
  seed_number = 1993,
  format = "habitat_viability"
)
}
\arguments{
\item{experiment}{Character. The path to the experiment directory containing cGENIE output files.}

\item{year}{Character or numeric. The year to process. If "default", the final time step is used.}

\item{n_ecotypes}{Integer. The number of ecophysiotypes (ecotypes) to model. Default is 1000.}

\item{seed_number}{Integer. The seed number for random number generation. Default is 1993.}

\item{format}{Character. The format of the output. Either "habitat_viability" or "ecotype_viability". Default is "habitat_viability".}
}
\value{
Depending on the format parameter, returns either a 3D array of habitat viability percentages or a 4D array of ecotype viability.
}
\description{
This function processes cGENIE model output to calculate habitat viability and ecotype viability based on metabolic index models.
}
\details{
This function reads cGENIE model output and calculates habitat viability and ecotype viability using metabolic index models. It sources functions from Hofmann et al. (2011) and uses parameter values from Penn et al. (2018). The function processes the data to generate arrays of partial pressure of oxygen (pO2) and temperature in Kelvin, and then calculates the viability of each ecotype in each model cell.
}
