% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.omz.map.R
\name{cGENIE.omz.map}
\alias{cGENIE.omz.map}
\title{Generate Maps from cGENIE Model Output}
\usage{
cGENIE.omz.map(
  experiment,
  dims = 3,
  thresh = 5.854e-05,
  year = "default",
  unit.factor = NULL,
  min.value = NULL,
  max.value = NULL,
  intervals = NULL,
  continents.outlined = TRUE,
  line.thickness = 1,
  scale.label = NULL,
  model = "biogem",
  palette_name = pals::parula(1000),
  projection = "ESRI:54012",
  darkmode = FALSE,
  bg.color = ifelse(darkmode, "black", "white"),
  fg.color = ifelse(darkmode, "white", "black")
)
}
\arguments{
\item{experiment}{(character) The path or name of the experiment used to locate the NetCDF file.}

\item{dims}{(numeric) The dimensionality of the data (default is 3 for 3D; can be 2D or 3D).}

\item{thresh}{A numeric value specifying the threshold oxygen concentration (in umol/kg) below which
the volume is considered part of the OMZ. Default is 58.54e-6 based on Canfield 60uM
https://enviromicro-journals.onlinelibrary.wiley.com/doi/10.1111/1462-2920.16192
(just a conversion based on global mean density of 1.025 kg/L)}

\item{year}{(numeric or character) Time step to visualize (default uses the final time step if "default").}

\item{unit.factor}{(numeric) A scaling factor for the variable values (default is 1).}

\item{min.value}{(numeric) Minimum value for the color scale (used to set scale limits).}

\item{max.value}{(numeric) Maximum value for the color scale.}

\item{intervals}{(numeric) Step intervals for the color scale.}

\item{continents.outlined}{(logical) Logical value to control whether to outline continents.}

\item{line.thickness}{(numeric) Thickness of the lines outlining the continents and other map elements.}

\item{scale.label}{(character) Label for the color bar.}

\item{model}{(character) The model type (default is 'biogem'; can be extended for other models).}

\item{palette_name}{(character) Color palette to be used for the plot (default is \code{pals::parula(1000)}).}

\item{projection}{(character) Map projection to use (default is ESRI:54012 for Equal Earth).}

\item{darkmode}{(logical) Logical value to control whether to use dark mode (default is FALSE).}

\item{bg.color}{(character) Background color for the plot (default is "black" when darkmode is TRUE).}

\item{fg.color}{(character) Foreground color for the plot (default is "white" when darkmode is TRUE).}
}
\value{
A ggplot object representing the generated map with the specified variable visualized across geographical coordinates.
}
\description{
This function generates a specific kind of maps from imported .nc (NetCDF) files containing data from the cGENIE model outputs.
It plots OMZ depths based on some threshold.
}
\details{
This function generates a map visualization of the global distribution of oxygen minimum zones from a given NetCDF file
containing data from the cGENIE model outputs.
}
