% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.omz.volume.R
\name{cGENIE.omz.volume}
\alias{cGENIE.omz.volume}
\title{Summarize the Volume of Oxygen Minimum Zones (OMZs)}
\usage{
cGENIE.omz.volume(experiment, thresh = 4.8e-06, time.step = "default")
}
\arguments{
\item{experiment}{A string representing the path to the experiment's directory containing the NetCDF files.}

\item{thresh}{A numeric value specifying the threshold oxygen concentration (in umol/kg) below which
the volume is considered part of the OMZ. Default is 4.8e-6.}

\item{time.step}{Either "default" to use the last time step or a numeric index to specify a particular time step.}
}
\value{
A numeric value representing the proportion of the ocean volume that is classified as an OMZ.
}
\description{
This function calculates the proportion of the ocean volume that is classified as oxygen minimum zones (OMZs)
based on the given threshold of oxygen concentration. The calculation is performed using data from
a specified experiment in NetCDF format.
}
\details{
The function extracts relevant variables from the NetCDF file, calculates the thickness of each depth layer,
and sums the volumes classified as OMZs and non-OMZs based on the provided threshold. The final output is
the proportion of the total volume that constitutes OMZs.
}
\examples{
# Calculate the OMZ volume proportion for a specific experiment
omz_volume <- cGENIE.omz.volume("path/to/experiment", thresh = 4.8e-6)

}
