% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.overlay.map.R
\name{cGENIE.overlay.map}
\alias{cGENIE.overlay.map}
\title{Overlay Map for cGENIE Data}
\usage{
cGENIE.overlay.map(
  var,
  experiment,
  depth.level = 1,
  dims = 3,
  year = "default",
  unit.factor = 1,
  min.value,
  max.value,
  intervals,
  continents.outlined = TRUE,
  scale.label,
  line.thickness = 1,
  model = "biogem",
  palette_name = pals::parula(1000),
  projection = "ESRI:54012",
  darkmode = FALSE,
  bg_color = ifelse(darkmode, "black", "white"),
  fg_color = ifelse(darkmode, "white", "black")
)
}
\arguments{
\item{var}{The variable name to extract from the .nc file.}

\item{experiment}{Directory containing the experiment's netCDF files.}

\item{depth.level}{The depth level to extract from the .nc file. Defaults to 1.}

\item{dims}{The number of dimensions for the data. Defaults to 3.}

\item{year}{The year to extract from the .nc file. Defaults to "default".}

\item{unit.factor}{The factor to multiply the variable values by. Defaults to 1.}

\item{min.value}{Minimum value for the color scale. Defaults to NULL.}

\item{max.value}{Maximum value for the color scale. Defaults to NULL.}

\item{intervals}{Number of intervals for the color scale. Defaults to NULL.}

\item{continents.outlined}{Logical indicating whether to outline continents. Defaults to TRUE.}

\item{scale.label}{Label for the color scale. Defaults to NULL.}

\item{line.thickness}{The thickness of the outline lines. Defaults to 1.}

\item{model}{The model type; defaults to "biogem".}

\item{palette_name}{The color palette to use for the map. Defaults to pals::parula(1000).}

\item{projection}{The map projection to use. Defaults to 'ESRI:54012'.}

\item{darkmode}{Logical indicating whether to use dark mode. Defaults to FALSE.}

\item{bg_color}{The background color of the map. Defaults to "white" if darkmode is FALSE, "black" if darkmode is TRUE.}

\item{fg_color}{The foreground color of the map. Defaults to "black" if darkmode is FALSE, "white" if darkmode is TRUE.}
}
\value{
A ggplot object representing the overlay map.
}
\description{
This function creates an overlay map for cGENIE data using various projections and dimensions.
}
