% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.point.matching.benthic.R
\name{cGENIE.point.matching.benthic}
\alias{cGENIE.point.matching.benthic}
\title{cGENIE Point Matching Benthic}
\usage{
cGENIE.point.matching.benthic(
  var = NULL,
  experiment = NULL,
  format = "nc",
  input = NULL,
  dims = 3,
  time.present = FALSE,
  coord.dat,
  lat.name = "p_lat",
  lng.name = "p_lng"
)
}
\arguments{
\item{var}{Character. The variable to extract from the cGENIE model. Not used if format is "array".}

\item{experiment}{Character. The experiment identifier for the cGENIE model. Not used if format is "array".}

\item{format}{Character. The format of the input data. Can be "nc" for netCDF or "array" for array format. Default is "nc".}

\item{input}{Character. The input data file name if the format is "array". Default is NULL. Not used if format is "nc".}

\item{dims}{Numeric. The number of dimensions in the cGENIE model. Default is 3.}

\item{time.present}{Logical. Whether to use the present time for the cGENIE model. Default is FALSE.}

\item{coord.dat}{Data frame. A data frame with latitude and longitude columns. cGENIE data will be added to this and returned.}

\item{lat.name}{Character. The name of the latitude column in \code{coord.dat}. Default is "p_lat".}

\item{lng.name}{Character. The name of the longitude column in \code{coord.dat}. Default is "p_lng".}
}
\value{
A data frame with the original coordinates and matched climate data from the cGENIE model.
}
\description{
This function matches palaeocoordinates to cGENIE model data.
}
