% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.points.matching.benthic.R
\name{cGENIE.points.matching.benthic}
\alias{cGENIE.points.matching.benthic}
\title{cGENIE Points Matching Benthic}
\usage{
cGENIE.points.matching.benthic(
  var = NULL,
  experiment = NULL,
  input = NULL,
  format = "nc",
  depth.level = 1,
  dims = 3,
  coord.dat = NULL,
  lat.name = "p_lat",
  lng.name = "p_lng",
  output = "dataframe"
)
}
\arguments{
\item{var}{Character. The variable to extract from the cGENIE model.}

\item{experiment}{Character. The experiment identifier for the cGENIE model.}

\item{input}{Array. The input array for the cGENIE model data when format is "array". Otherwise does not need to be specified.}

\item{format}{Character. The format of the input data. Default is "nc". Options are "nc" or "array".}

\item{depth.level}{Numeric. The depth level to extract from the cGENIE model. Default is 1.}

\item{dims}{Numeric. The number of dimensions in the cGENIE model. Default is 3.}

\item{coord.dat}{Data frame. A data frame with latitude and longitude columns. cGENIE data will be added to this and returned.}

\item{lat.name}{Character. The name of the latitude column in \code{coord.dat}. Default is "p_lat".}

\item{lng.name}{Character. The name of the longitude column in \code{coord.dat}. Default is "p_lng".}

\item{output}{Character. The format of the output. Default is "dataframe". Options are "dataframe" or "array".}
}
\value{
A data frame or 3D array with the original coordinates and matched climate data from the cGENIE model.
}
\description{
This function matches palaeocoordinates to cGENIE model data for benthic points.
}
