% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.res.final.R
\name{cGENIE.res.final}
\alias{cGENIE.res.final}
\title{cGENIE.res.final}
\usage{
cGENIE.res.final(var, sub_var = "default", experiment)
}
\arguments{
\item{var}{A character string specifying the variable to be imported. Options include "ocn_temp" and "ocn_O2".}

\item{sub_var}{A character string specifying the sub-variable to be used. Default is "default".}

\item{experiment}{A character string specifying the experiment from which the data is to be imported.}
}
\value{
The final data point of the specified variable from the imported .res file.
}
\description{
This function imports .res files and converts them into data frames, then extracts the final data point.
}
\details{
The function first imports the .res file corresponding to the specified variable and experiment using the \code{cGENIE.res.import} function.
It then extracts the final data point of the specified variable. If \code{sub_var} is set to "default", the function uses predefined sub-variables
for "ocn_temp" and "ocn_O2". For other variables, it extracts the value from the second column of the final row.
}
\examples{
\dontrun{
final_temp <- cGENIE.res.final(var = "ocn_temp", experiment = "experiment1")
final_O2 <- cGENIE.res.final(var = "ocn_O2", experiment = "experiment2")
}

}
