% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.shelf.R
\name{cGENIE.shelf}
\alias{cGENIE.shelf}
\title{Summarize Shelf Environments from cGENIE .nc Files}
\usage{
cGENIE.shelf(
  input,
  format = "array",
  shelf.depth = 3,
  array.only = TRUE,
  surf = FALSE,
  time.step = "default"
)
}
\arguments{
\item{format}{A character string specifying the format of the output. Default is "array".}

\item{shelf.depth}{An integer specifying the number of vertical layers to consider as the shelf. Default is 3.}

\item{array.only}{A logical value indicating whether to return only the array of the variable for shelf synthesis. Default is TRUE.}

\item{surf}{A logical value indicating whether to include the surface layer in the shelf. Default is FALSE.}

\item{time.step}{A character or integer specifying the time step to use. Default is "default", which uses the length of \code{nc.sum$time}.}

\item{nc.sum}{A list containing the NetCDF summary data, including variables such as \code{var}, \code{time}, \code{lon}, \code{lat}, and \code{depth}.}
}
\value{
An array summarizing the shelf environments.
}
\description{
This function summarizes the shelf environments from cGENIE .nc files.
The default shelf is the same as that used in Stockey et al. 2021 PNAS and Pohl et al. 2023 Sci Adv,
i.e., the top 3 layers of cells directly adjacent to land. The function can be customized to alter
the vertical number of cells and the horizontal distance from land.
}
