% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.sum.plots.R
\name{cGENIE.sum.plots}
\alias{cGENIE.sum.plots}
\title{Generate Key Summary Plots for cGENIE Models}
\usage{
cGENIE.sum.plots(
  experiment = NULL,
  directory = "default",
  file.name = "cGENIE.sum.plots",
  format = "pdf",
  save = FALSE
)
}
\arguments{
\item{experiment}{A character string specifying the experiment name.}

\item{directory}{A character string specifying the directory to save the plot. Default is the current working directory.}

\item{file.name}{A character string specifying the name of the output file. Default is "cGENIE.sum.plots".}

\item{format}{A character string specifying the format of the output file. Default is "pdf".}

\item{save}{A logical value indicating whether to save the plot. Default is FALSE.}
}
\value{
A ggplot object containing the arranged summary plots.
}
\description{
This function generates a set of key summary plots for cGENIE models,
designed to fit on one A4-ish sheet of paper. The plots include:
\itemize{
\item Sea surface temperature through time
\item Sea surface temperature map
\item Bottom water temperature map
\item Global marine O2 through time
\item Second level marine O2 map
\item Bottom water O2 map
}
}
\details{
The function produces a 3x2 landscape paneled plot.
}
\examples{
\dontrun{
cGENIE.sum.plots(experiment = "experiment_name", save = TRUE)
}

}
