% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.trouble.plots.simple.R
\name{cGENIE.trouble.plots.simple}
\alias{cGENIE.trouble.plots.simple}
\title{cGENIE.trouble.plots}
\usage{
cGENIE.trouble.plots.simple(
  experiment,
  directory = "default",
  file.name = "cGENIE.trouble.plots",
  format = "pdf",
  save = FALSE
)
}
\arguments{
\item{experiment}{The experiment data to be used for generating the plots.}

\item{directory}{The directory where the plot will be saved. Default is the current working directory.}

\item{file.name}{The name of the file to save the plot as. Default is "cGENIE.trouble.plots".}

\item{format}{The format of the saved plot file. Default is "pdf".}

\item{save}{Logical indicating whether to save the plot. Default is FALSE.}
}
\value{
A ggplot object containing the arranged summary plots.
}
\description{
Generates key summary plots for cGENIE models on one A4-ish sheet of paper. The key plots include:
\itemize{
\item Sea surface temperature through time
\item Bathymetry
\item Barotropic streamfunction
\item Global marine O2 through time
\item Surface alkalinity map (commented out)
\item Surface H2S map (commented out)
}
}
\details{
This function produces a 3x2 landscape paneled plot.
}
