% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.continents.sf.R
\name{gen.continents.sf}
\alias{gen.continents.sf}
\title{Generate continental outlines from cleaned model output}
\usage{
gen.continents.sf(df)
}
\arguments{
\item{df}{Dataframe from \code{clean.nc.df()} containing at least columns:
lon.min, lon.max, lat.min, lat.max, var}
}
\value{
An \code{sf} object representing continental outlines
}
\description{
Converts grid cells with NA values (assumed land) into a single
\code{sf} object representing continental outlines.
}
