% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HADCM3.benthic.data.R
\name{HADCM3.benthic.data}
\alias{HADCM3.benthic.data}
\title{Extract Benthic Data from NetCDF Files}
\usage{
HADCM3.benthic.data(var, file, experiment, time.present = FALSE, na.rm = FALSE)
}
\arguments{
\item{var}{Character. The variable to extract from the NetCDF file.}

\item{file}{Character. The path to the NetCDF file.}

\item{experiment}{Character. The experiment identifier.}

\item{time.present}{Logical. If TRUE, extracts data for the present time. Default is FALSE.}

\item{na.rm}{Logical. If TRUE, removes NA values. Default is FALSE.}
}
\value{
A data frame containing the extracted benthic data.
}
\description{
This function extracts benthic data from imported NetCDF (.nc) files, such as those from Valdes et al. (2021).
}
\details{
This function reads in a NetCDF file and extracts data for benthic (seafloor) cells. It uses several R packages for data manipulation and visualization, including \code{RNetCDF}, \code{dplyr}, \code{sf}, \code{sp}, \code{ggspatial}, \code{reshape2}, \code{ggplot2}, \code{pals}, and \code{viridis}.

The function first calls \code{HADCM3.whole.ocean.data} to get the full dataset, then filters this dataset to retain only the seafloor cells. The resulting data frame is returned.
}
\note{
Other projection options include Lambert Cylindrical Equal Area (EPSG: 6933). The \code{palette_name} parameter currently requires a numeric value (e.g., \code{parula(1000)}). Other color palette options include \code{viridis} and others available at \url{https://r-charts.com/color-palettes/}.
}
