% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HADCM3.data.3D.R
\name{HADCM3.data.3D}
\alias{HADCM3.data.3D}
\title{Extract and Process 3D Data from HADCM3 NetCDF Files}
\usage{
HADCM3.data.3D(var, file, experiment, year = "default")
}
\arguments{
\item{var}{A character string specifying the name of the variable to extract from the NetCDF file.}

\item{file}{A character string indicating the name of the NetCDF file (without the .nc extension).}

\item{experiment}{A character string indicating the path to the HADCM3 experiment directory.}

\item{year}{Year to extract data for (default is "default", meaning the last or only time point (note this is slightly different to cGENIE.data because HADCM3 outputs more commonly have only one time point)).}
}
\value{
A dataframe (\code{df.sum}) containing:
\itemize{
\item \code{lon.mid}: Midpoint of the longitude.
\item \code{lon.min}: Minimum longitude boundary.
\item \code{lon.max}: Maximum longitude boundary.
\item \code{lat.mid}: Midpoint of the latitude.
\item \code{lat.min}: Minimum latitude boundary.
\item \code{lat.max}: Maximum latitude boundary.
\item \code{var}: Extracted variable values (e.g., temperature, salinity).
\item \code{depth.level}: Depth level index (1 for surface, increasing with depth).
\item \code{depth}: Actual depth value in meters.
\item \code{depth.min}: Minimum depth boundary.
\item \code{depth.max}: Maximum depth boundary.
\item \code{lon.range}: Range of longitude for handling map boundaries.
}
}
\description{
This function extracts and processes 3D data from HADCM3 model NetCDF files
based on latitude, longitude, depth, and time. It converts the 3D array into a
2D dataframe with geographical coordinates and depth levels, making the data
easier to visualize or further process.
}
