% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HADCM3.data.R
\name{HADCM3.data}
\alias{HADCM3.data}
\title{Extract Data from HADCM3 NetCDF Files}
\usage{
HADCM3.data(var, file, experiment, depth.level = 1, dims = 3, year = "default")
}
\arguments{
\item{var}{A character string specifying the name of the variable to extract from the NetCDF file.}

\item{file}{A character string indicating the name of the NetCDF file (without the .nc extension).}

\item{experiment}{A character string indicating the path to the HADCM3 experiment directory.}

\item{depth.level}{Depth level to extract the data from. Set to NULL by default.}

\item{dims}{An integer specifying the dimensions of the NetCDF file to read. Default is set to 3 for 3D data but can alternatively be set to 2. .}

\item{year}{Year to extract data for (default is "default", meaning the last or only time point (note this is slightly different to cGENIE.data because HADCM3 outputs more commonly have only one time point)).}
}
\value{
A data frame containing the extracted 2D lat-lon data field with corresponding grid information.
}
\description{
This function extracts data from imported NetCDF (.nc) files, specifically designed for use with HADCM3 model outputs.
}
