% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HADCM3.get.nc.R
\name{HADCM3.get.nc}
\alias{HADCM3.get.nc}
\title{Import Data from HADCM3 NetCDF Files}
\usage{
HADCM3.get.nc(var, file, experiment, dims = 3)
}
\arguments{
\item{var}{A character string specifying the name of the variable to extract from the NetCDF file.}

\item{file}{A character string indicating the name of the NetCDF file (without the .nc extension).}

\item{experiment}{A character string indicating the path to the HADCM3 experiment directory.}

\item{dims}{An integer specifying the dimensions of the NetCDF file to read. Default is set to 3 for 3D data but can alternatively be set to 2. .}
}
\value{
A list containing latitude, longitude, depth, time, and the specified variable data from the NetCDF file.
}
\description{
This function imports specified variables from HADCM3 NetCDF files (as formatted by the BRIDGE Group) and extracts relevant grid information.
}
\details{
The function reads a NetCDF file based on the specified experiment and model, extracting latitude, longitude,
depth, time, and the requested variable. It supports both 2D and 3D datasets.
}
