% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HADCM3.m2.to.global.R
\name{HADCM3.m2.to.global}
\alias{HADCM3.m2.to.global}
\title{HADCM3.m2.to.global}
\usage{
HADCM3.m2.to.global(
  var,
  file,
  experiment,
  depth.level = 1,
  dims = 2,
  unit.factor = 1,
  time.present = FALSE
)
}
\arguments{
\item{var}{A character string specifying the variable to be extracted from the NetCDF file.}

\item{file}{A character string specifying the file name (without extension) to be read from the experiment directory.}

\item{experiment}{A character string specifying the path to the experiment directory.}

\item{depth.level}{An integer specifying the depth level to be used for 3D data. Default is 1.}

\item{dims}{An integer specifying the dimensions of the NetCDF data (2 or 3). Default is 2.}

\item{unit.factor}{A numeric value to scale the variable. Default is 1.}

\item{time.present}{A logical value indicating if time data is present in the NetCDF file. Default is FALSE.}
}
\value{
A numeric value representing the total global value of the specified variable.
}
\description{
This function upscales per area estimates of variables (e.g., triffid outputs) to global values using data from HADCM3 model outputs.
}
\details{
The function reads NetCDF files using the RNetCDF package and extracts latitude, longitude, and optionally depth and time data. It then processes the data to create a dataframe with the specified variable and its corresponding spatial coordinates. The function calculates the area of each grid cell and multiplies it by the variable value to obtain the total global value.
}
\note{
\itemize{
\item The function assumes that the NetCDF files follow a specific structure and naming convention.
\item The function includes several checks and adjustments for the HADCM3 grid geometry.
\item The function uses various R packages for data manipulation and spatial calculations, including dplyr, sf, sp, ggspatial, reshape2, ggplot2, pals, viridis, and geosphere.
}
}
\examples{
\dontrun{
# Example usage:
result <- HADCM3.m2.to.global(var = "insitu_T_ym_dpth",
                              file = "o.pgclann",
                              experiment = "~/Valdes2021_HADCM3L/teXPl_444/teXPl_444",
                              depth.level = 1,
                              dims = 3,
                              unit.factor = 1,
                              time.present = FALSE)
print(result)
}

}
