% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HADCM3.map.R
\name{HADCM3.map}
\alias{HADCM3.map}
\title{Generate Maps from HADCM3 Model Output (as formatted by the BRIDGE Group)}
\usage{
HADCM3.map(
  var,
  file,
  experiment,
  depth.level = 1,
  dims = NULL,
  year = "default",
  unit.factor = NULL,
  min.value = NULL,
  max.value = NULL,
  intervals = NULL,
  continents.outlined = TRUE,
  scale.label = NULL,
  line.thickness = 1,
  palette_name = pals::parula(1000),
  projection = "ESRI:54012",
  darkmode = FALSE,
  custom.bg = NULL,
  custom.fg = NULL,
  col.labels = NULL
)
}
\arguments{
\item{var}{A character string specifying the name of the variable to extract from the NetCDF file. (e.g. "insitu_T_ym_dpth")}

\item{file}{A character string indicating the name of the NetCDF file (without the .nc extension).}

\item{experiment}{A character string indicating the path to the HADCM3 experiment directory.}

\item{depth.level}{(numeric) Depth layer to visualize (default = 1).}

\item{dims}{(numeric) Dimensionality of the data (default NULL; auto-set by var).}

\item{year}{(numeric or character) Time step to visualize (default "default").}

\item{unit.factor}{(numeric) Scaling factor for variable (default NULL; auto-set by var).}

\item{min.value, max.value}{(numeric) Color scale limits (default NULL; auto-set by var).}

\item{intervals}{(numeric) Color scale intervals (default NULL; auto-set by var).}

\item{continents.outlined}{(logical) Whether to outline continents (default TRUE).}

\item{scale.label}{(character) Label for color scale (default NULL; auto-set by var).}

\item{line.thickness}{(numeric) Thickness of continent outlines (default 1).}

\item{palette_name}{(character) Color palette for the plot (default pals::parula(1000)).}

\item{projection}{(character) Map projection (default 'ESRI:54012').}

\item{darkmode}{(logical) Enable dark mode (default FALSE).}

\item{custom.bg}{(character) Background color for dark mode (default NULL; auto-set if darkmode = TRUE).}

\item{custom.fg}{(character) Foreground color for dark mode (default NULL; auto-set if darkmode = TRUE).}

\item{col.labels}{(numeric) Labels for the color bar ticks.}
}
\value{
A ggplot object representing the map.
}
\description{
Generate Maps from HADCM3 Model Output (as formatted by the BRIDGE Group)
}
