% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HADCM3.point.matching.benthic.R
\name{HADCM3.point.matching.benthic}
\alias{HADCM3.point.matching.benthic}
\title{HADCM3 Point Matching Benthic}
\usage{
HADCM3.point.matching.benthic(
  var = NULL,
  file = NULL,
  experiment = NULL,
  dims = 3,
  time.present = FALSE,
  coord.dat,
  lat.name = "p_lat",
  lng.name = "p_lng"
)
}
\arguments{
\item{var}{Character. The variable to extract from the climate data file. Default is NULL.}

\item{file}{Character. The file name of the cGENIE netCDF file. Default is NULL.}

\item{experiment}{Character. The experiment name in the cGENIE netCDF file. Default is NULL.}

\item{dims}{Numeric. The number of dimensions in the climate data file. Default is 3.}

\item{time.present}{Logical. Whether the time is present in the climate data file. Default is FALSE.}

\item{coord.dat}{Data frame. A data frame with the latitude and longitude column names assigned. cGENIE data will be added to this and returned.}

\item{lat.name}{Character. The name of the latitude column in the coord.dat data frame. Default is "p_lat".}

\item{lng.name}{Character. The name of the longitude column in the coord.dat data frame. Default is "p_lng".}
}
\value{
A data frame with the matched climate data added.
}
\description{
This function matches palaeocoordinates to cGENIE models by extracting grid and climate data from cGENIE netCDF files and matching them to the provided coordinates.
This version of the function is specifically designed for benthic data.
}
