% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HADCM3.point.matching.R
\name{HADCM3.point.matching}
\alias{HADCM3.point.matching}
\title{HADCM3 Point Matching}
\usage{
HADCM3.point.matching(
  var = NULL,
  file = NULL,
  experiment = NULL,
  depth.level = 1,
  dims = 3,
  time.present = FALSE,
  coord.dat,
  lat.name = "p_lat",
  lng.name = "p_lng"
)
}
\arguments{
\item{var}{Character. The variable to extract from the climate data file. Default is NULL.}

\item{file}{Character. The file name of the cGENIE netCDF file. Default is NULL.}

\item{experiment}{Character. The experiment name in the cGENIE netCDF file. Default is NULL.}

\item{depth.level}{Numeric. The depth level to extract from the climate data file. Default is 1.}

\item{dims}{Numeric. The number of dimensions in the climate data file. Default is 3.}

\item{time.present}{Logical. Whether the time is present in the climate data file. Default is FALSE.}

\item{coord.dat}{Data frame. A data frame with the latitude and longitude column names assigned. cGENIE data will be added to this and returned.}

\item{lat.name}{Character. The name of the latitude column in the coord.dat data frame. Default is "p_lat".}

\item{lng.name}{Character. The name of the longitude column in the coord.dat data frame. Default is "p_lng".}
}
\value{
A data frame with the matched climate data added.
}
\description{
This function matches palaeocoordinates to cGENIE models by extracting grid and climate data from cGENIE netCDF files and matching them to the provided coordinates.
}
\examples{
\dontrun{
coord.dat <- data.frame(p_lat = c(10, 20), p_lng = c(30, 40))
result <- HADCM3.point.matching(var = "insitu_T_ym_dpth", file = "o.pgclann", experiment = "experiment1", coord.dat = coord.dat)
}

}
