% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HADCM3.points.map.R
\name{HADCM3.points.map}
\alias{HADCM3.points.map}
\title{Generate Maps from HADCM3L Model Output with Points Matching}
\usage{
HADCM3.points.map(
  var,
  file,
  experiment,
  depth.level = 1,
  dims = NULL,
  year = "default",
  coord.dat,
  lat.name = "p_lat",
  lng.name = "p_lng",
  max_dist = 1000,
  na.method = "nearest",
  unit.factor = NULL,
  min.value = NULL,
  max.value = NULL,
  intervals = NULL,
  continents.outlined = TRUE,
  scale.label = NULL,
  line.thickness = 1,
  palette_name = pals::parula(1000),
  projection = "ESRI:54012",
  darkmode = FALSE,
  custom.bg = NULL,
  custom.fg = NULL,
  col.labels = NULL,
  show.NAs = TRUE
)
}
\arguments{
\item{var}{A character string specifying the name of the variable to extract from the NetCDF file. (e.g. "insitu_T_ym_dpth")}

\item{file}{A character string indicating the name of the NetCDF file (without the .nc extension).}

\item{experiment}{A character string indicating the path to the HADCM3 experiment directory.}

\item{depth.level}{(numeric) Depth layer to visualize (default = 1).}

\item{dims}{(numeric) Dimensionality of the data (default NULL).}

\item{year}{(numeric or character) Time step to visualize (default "default").}

\item{coord.dat}{Data frame. A data frame with latitude and longitude columns. cGENIE data will be added to this and returned.}

\item{lat.name}{Character. The name of the latitude column in \code{coord.dat}. Default is "p_lat".}

\item{lng.name}{Character. The name of the longitude column in \code{coord.dat}. Default is "p_lng".}

\item{max_dist}{Maximum allowed distance for matching (km, default 1000 km).}

\item{na.method}{How to handle NA climate values: "nearest" (default), "same.lat", "keep".}

\item{unit.factor}{(numeric) Scaling factor for variable (default NULL).}

\item{min.value, max.value}{(numeric) Color scale limits (default NULL).}

\item{intervals}{(numeric) Color scale intervals (default NULL).}

\item{continents.outlined}{(logical) Whether to outline continents (default TRUE).}

\item{scale.label}{(character) Label for color scale (default NULL).}

\item{line.thickness}{(numeric) Thickness of continent outlines (default 1).}

\item{palette_name}{(character) Color palette for the plot (default pals::parula(1000)).}

\item{projection}{(character) Map projection (default 'ESRI:54012').}

\item{darkmode}{(logical) Enable dark mode (default FALSE).}

\item{custom.bg}{(character) Background color for dark mode (default NULL; auto-set if darkmode = TRUE).}

\item{custom.fg}{(character) Foreground color for dark mode (default NULL; auto-set if darkmode = TRUE).}

\item{col.labels}{(numeric) Labels for the color bar ticks.}

\item{show.NAs}{(logical) Whether to show NA values on the map (default FALSE).}
}
\value{
A ggplot object representing the map with matched points.
}
\description{
This function generates maps from imported HADCM3L .nc files. It can handle both 2D and 3D netCDF files and provides options for various customizations including projections, color palettes, and more.
Additionally, it matches and plots specific points from a provided data frame.
}
\details{
This function reads 2D or 3D data from HADCM3L model output NetCDF files and produces a map visualization.
Users can specify their own scaling and color settings.
The function also matches specific points from a provided data frame and plots them on the map.
}
