% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HADCM3.transect.R
\name{HADCM3.transect}
\alias{HADCM3.transect}
\title{Create Transect Plots from HADCM3 Model Output (as formatted by the BRIDGE Group)}
\usage{
HADCM3.transect(
  var,
  file,
  experiment,
  degrees = 0,
  dims = 3,
  orientation = "lat",
  year = "default",
  unit.factor = 1,
  min.value = 0,
  max.value = 40,
  intervals = 5,
  continents.outlined,
  scale.label = NULL,
  line.thickness = 1,
  palette_name = pals::parula(1000),
  projection = "ESRI:54012",
  darkmode = FALSE,
  custom.bg = NULL,
  custom.fg = NULL,
  col.labels = NULL,
  upper.bound = -20,
  lower.bound = "default"
)
}
\arguments{
\item{var}{A character string specifying the name of the variable to extract from the NetCDF file. (e.g. "insitu_T_ym_dpth")}

\item{file}{A character string indicating the name of the NetCDF file (without the .nc extension).}

\item{experiment}{A character string indicating the path to the HADCM3 experiment directory.}

\item{degrees}{Numeric. The degree band of the specified latitude or longitude to include in the transect (default 0, meaning the equator/dateline). Will plot the nearest band to your chosen value.}

\item{dims}{(numeric) Dimensionality of the data (almost certainly 3 if you're plotting a transect, set to 3 as default).}

\item{orientation}{Character. Whether the transect is oriented along latitude ("lat") or longitude ("lon"). Default is "lat".}

\item{year}{(numeric or character) Time step to visualize (default "default").}

\item{unit.factor}{(numeric) Scaling factor for variable (default NULL; auto-set by var).}

\item{min.value, max.value}{(numeric) Color scale limits (default NULL; auto-set by var).}

\item{intervals}{(numeric) Color scale intervals (default NULL; auto-set by var).}

\item{continents.outlined}{(logical) Whether to outline continents (default TRUE).}

\item{scale.label}{(character) Label for color scale (default NULL; auto-set by var).}

\item{line.thickness}{(numeric) Thickness of continent outlines (default 1).}

\item{palette_name}{(character) Color palette for the plot (default pals::parula(1000)).}

\item{projection}{(character) Map projection (default 'ESRI:54012').}

\item{darkmode}{(logical) Enable dark mode (default FALSE).}

\item{custom.bg}{(character) Background color for dark mode (default NULL; auto-set if darkmode = TRUE).}

\item{custom.fg}{(character) Foreground color for dark mode (default NULL; auto-set if darkmode = TRUE).}

\item{col.labels}{(numeric) Labels for the color bar ticks.}

\item{upper.bound}{Numeric value indicating the upper bound for the y axis. Defaults to -20 to avoid floating y axis ticks. Can be tweaked for aesthetics.}

\item{lower.bound}{Numeric value indicating the lower bound for the y axis. Defaults to "default", which sets it to the maximum depth in the data.}
}
\value{
A ggplot object representing the transect plot.
}
\description{
This function creates transect plots from HADCM3 model output .nc files.
}
