% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HADCM3.whole.ocean.data.R
\name{HADCM3.whole.ocean.data}
\alias{HADCM3.whole.ocean.data}
\title{Extract Data from HADCM3 NetCDF Files}
\usage{
HADCM3.whole.ocean.data(
  var,
  file,
  experiment,
  time.present = FALSE,
  na.rm = FALSE
)
}
\arguments{
\item{var}{Character. The name of the variable to extract from the NetCDF file.}

\item{file}{Character. The name of the NetCDF file (without extension) to read from.}

\item{experiment}{Character. The path to the experiment directory containing the NetCDF file.}

\item{time.present}{Logical. If TRUE, extracts the time variable from the NetCDF file. Default is FALSE.}

\item{na.rm}{Logical. If TRUE, removes rows with NA values from the resulting dataframe. Default is FALSE.}
}
\value{
A dataframe with the following columns:
\describe{
\item{lon.mid}{Numeric. The mid-point longitude values.}
\item{lon.min}{Numeric. The minimum longitude values.}
\item{lon.max}{Numeric. The maximum longitude values.}
\item{lat.mid}{Numeric. The mid-point latitude values.}
\item{lat.min}{Numeric. The minimum latitude values.}
\item{lat.max}{Numeric. The maximum latitude values.}
\item{depth.level}{Numeric. The depth level index.}
\item{var}{Numeric. The extracted variable values.}
}
}
\description{
This function extracts data from imported NetCDF (.nc) files, specifically designed for the HADCM3 model.
It processes the data to generate a dataframe suitable for further analysis and visualization.
}
\details{
This function reads a 3D array from the specified NetCDF file and processes it into a 2D dataframe.
It adjusts the longitude values to be within the range of -180 to 180 degrees if necessary.
The function also handles the extraction of latitude, longitude, and depth variables, and can optionally
extract the time variable if \code{time.present} is set to TRUE.
}
\note{
The function assumes that the NetCDF file follows the HADCM3 grid conventions. It may require adjustments
for other grid configurations.
}
\examples{
\dontrun{
# Example usage:
df <- HADCM3.whole.ocean.data(var = "insitu_T_ym_dpth", file = "o.pgclann", experiment = "~/Valdes2021_HADCM3L/teXPl_444/teXPl_444")
}

}
