% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleo.transect.R
\name{paleo.transect}
\alias{paleo.transect}
\title{Plot a Transect from Gridded Data}
\usage{
paleo.transect(
  df.slice,
  orientation = "lat",
  box,
  land_segments,
  unit.factor = 1,
  palette_name,
  min.value,
  max.value,
  intervals,
  custom.fg,
  upper.bound,
  scale.label,
  title_text,
  line.thickness
)
}
\arguments{
\item{df.slice}{Filtered dataframe for the transect.}

\item{orientation}{"lat" or "lon".}

\item{box}{Dataframe defining plot bounding box.}

\item{land_segments}{Dataframe of land outline segments.}

\item{unit.factor}{Numeric multiplier for variable scaling.}

\item{palette_name}{Colour palette.}

\item{min.value}{Minimum colour scale value.}

\item{max.value}{Maximum colour scale value.}

\item{intervals}{Interval for colour breaks.}

\item{custom.fg}{Foreground colour.}

\item{upper.bound}{Upper y-axis bound.}

\item{scale.label}{Legend label.}

\item{title_text}{Plot title.}

\item{line.thickness}{Thickness of land outline and box borders. (land outline half of box borders by default)}
}
\value{
ggplot object
}
\description{
Utility function to create a latitude or longitude transect plot
from pre-filtered gridded data (e.g. using gen.transect.slice).
}
