% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scotese.matrix.R
\name{scotese.matrix}
\alias{scotese.matrix}
\title{Generate a Matrix from a NetCDF File}
\usage{
scotese.matrix(
  map,
  min.value = -6000,
  max.value = 6000,
  intervals = 2000,
  continents.outlined,
  scale.label,
  res = "deg",
  scale = "viridis",
  projection = "ESRI:54012"
)
}
\arguments{
\item{map}{Character. Path to the NetCDF file.}

\item{min.value}{Numeric. Minimum value for the matrix. Default is -6000.}

\item{max.value}{Numeric. Maximum value for the matrix. Default is 6000.}

\item{intervals}{Numeric. Intervals for the matrix. Default is 2000.}

\item{continents.outlined}{Logical. Whether to outline continents. No default value provided.}

\item{scale.label}{Character. Label for the scale. No default value provided.}

\item{res}{Character. Resolution of the NetCDF file. Options are "deg" for degrees or "6min" for 6 minutes. Default is "deg".}

\item{scale}{Character. Color scale to use. Default is "viridis".}

\item{projection}{Character. Projection to use for the matrix. Default is 'ESRI:54012'.}
}
\value{
A matrix of values extracted from the NetCDF file.
}
\description{
This function reads a NetCDF file and generates a matrix based on the specified parameters.
}
\details{
This function supports different projections and resolutions for the NetCDF file. It extracts latitude and longitude
values and computes edges for these values. The main variable extracted from the NetCDF file is "z", which represents
elevation in meters.
}
\examples{
\dontrun{
  z_matrix <- scotese.matrix(
    map = "path/to/netcdf/file.nc",
    min.value = -6000,
    max.value = 6000,
    intervals = 2000,
    continents.outlined = TRUE,
    scale.label = "Elevation",
    res = "deg",
    scale = "viridis",
    projection = 'ESRI:54012'
  )
}

}
