% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.grid.R
\name{cGENIE.grid}
\alias{cGENIE.grid}
\title{Extract Grid Data from cGENIE NetCDF Files}
\usage{
cGENIE.grid(experiment, dims, model = "biogem")
}
\arguments{
\item{experiment}{A character string indicating the path to the cGENIE experiment directory.}

\item{dims}{An integer specifying the dimensions of the NetCDF file to read. Default is set to 3 for 3D data.}

\item{model}{A character string indicating the model type (default is "biogem"). Currently, only "biogem" is supported.}
}
\value{
A list containing latitude, longitude, and depth information, along with their respective edges.
}
\description{
This function extracts grid-related data (latitude, longitude, and depth) from a specified cGENIE NetCDF file.
It can handle both 2D and 3D grid data based on the specified dimensions.
}
\details{
The function handles the adjustment of longitude values to ensure they are within the range of \link{-180, 180}.
It returns a list containing the relevant grid data based on the specified dimensions (2D or 3D).
}
\examples{
# Example usage:
grid_data <- cGENIE.grid("experiment_directory", dims = 3)
}
