% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.grid.R
\name{cGENIE.grid}
\alias{cGENIE.grid}
\title{Extract Grid Data from cGENIE NetCDF Files}
\usage{
cGENIE.grid(experiment = NULL, dims = NULL, model = "biogem")
}
\arguments{
\item{experiment}{Character string indicating the path to the cGENIE experiment directory.}

\item{dims}{Integer specifying the dimensions of the NetCDF file to read. Default is NULL (returns default grid).}

\item{model}{Character string indicating the model type (default "biogem"). Currently only "biogem" is supported.}
}
\value{
A list containing latitude, longitude, depth (if dims=3), and their respective edges.
}
\description{
This function extracts grid-related data (latitude, longitude, and depth)
from a specified cGENIE NetCDF file. It handles both 2D and 3D grid data
based on the specified dimensions. If \code{experiment} or \code{dims} is \code{NULL},
it returns the default 36x36x16 cGENIE grid dimensions.
}
\details{
The function adjusts longitude values to be within -180 to 180 and reports if any values were outside this range.
}
