% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.overlay.map.R
\name{cGENIE.overlay.map}
\alias{cGENIE.overlay.map}
\title{Overlay Map for cGENIE Data}
\usage{
cGENIE.overlay.map(
  var,
  experiment,
  depth.level = 1,
  dims = 3,
  year = "default",
  unit.factor = 1,
  min.value,
  max.value,
  intervals,
  continents.outlined,
  scale.label,
  model = "biogem",
  palette_name = pals::parula(1000),
  projection = "ESRI:54012"
)
}
\arguments{
\item{var}{The variable name to extract from the .nc file.}

\item{experiment}{Directory containing the experiment's netCDF files.}

\item{min.value}{Minimum value for the color scale. Defaults to NULL.}

\item{max.value}{Maximum value for the color scale. Defaults to NULL.}

\item{intervals}{Number of intervals for the color scale. Defaults to NULL.}

\item{scale.label}{Label for the color scale. Defaults to NULL.}

\item{model}{The model type; defaults to "biogem".}
}
\value{
A ggplot object representing the overlay map.
}
\description{
This function creates an overlay map for cGENIE data using various projections and dimensions.
}
\examples{
\dontrun{
overlay_map <- cGENIE.overlay.map(var = "temperature", experiment = "experiment1")
}
}
