% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.res.plot.R
\name{cGENIE.res.plot}
\alias{cGENIE.res.plot}
\title{cGENIE.res.plot}
\usage{
cGENIE.res.plot(var, sub_var = "default", experiment)
}
\arguments{
\item{var}{A character string specifying the variable to be plotted (e.g., "ocn_temp", "ocn_O2", "atm_pO2", "atm_pCO2").}

\item{sub_var}{A character string specifying the sub-variable to be plotted. Default is "default".}

\item{experiment}{A character string specifying the experiment name.}
}
\value{
A ggplot object representing the time series plot of the specified variable.
}
\description{
This function imports .res files and converts them into data frames, then plots them as time series.
It builds on the \code{cGENIE.res.import} function.
}
\examples{
\dontrun{
plot <- cGENIE.res.plot(var = "ocn_temp", experiment = "experiment1")
print(plot)
}

}
