% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.transect.R
\name{cGENIE.transect}
\alias{cGENIE.transect}
\title{Create Transect Plots from cGENIE Model Output}
\usage{
cGENIE.transect(
  experiment,
  var,
  degrees = 0,
  dims = 3,
  orientation = "lat",
  year = "default",
  unit.factor = 1,
  min.value = 0,
  max.value = 40,
  intervals = 5,
  continents.outlined,
  scale.label = NULL,
  model = "biogem",
  palette_name = pals::parula(1000),
  projection = "ESRI:54012",
  darkmode = FALSE,
  custom.bg = NULL,
  custom.fg = NULL,
  upper.bound = -20,
  lower.bound = "default"
)
}
\arguments{
\item{experiment}{Directory containing the experiment's netCDF files.}

\item{var}{The variable name to extract from the .nc file.}

\item{degrees}{The latitude or longitude at which to extract from the data. Defaults to 0.}

\item{dims}{Number of dimensions in the netCDF file. Defaults to 3.}

\item{orientation}{Whether to create a transect along latitude or longitude. Defaults to "lat".}

\item{year}{Year to extract data for (default is "default", meaning the last time point).}

\item{unit.factor}{Factor to convert units. Defaults to 1.}

\item{min.value}{Minimum value for the color scale. Defaults to 0.}

\item{max.value}{Maximum value for the color scale. Defaults to 40.}

\item{intervals}{Number of intervals for the color scale. Defaults to 5.}

\item{continents.outlined}{Logical indicating whether to outline continents. Defaults to NULL.}

\item{scale.label}{Label for the color scale. Defaults to NULL.}

\item{model}{The model type; defaults to "biogem".}

\item{palette_name}{Color palette to use for the plot. Defaults to "pals::parula(1000)".}

\item{projection}{Projection to use for the plot. Defaults to "ESRI:54012".}

\item{darkmode}{Logical indicating whether to use a dark mode color scheme. Defaults to FALSE.}

\item{custom.bg}{Custom background color for dark mode. Defaults to NULL (auto-set if darkmode = TRUE).}

\item{custom.fg}{Custom foreground color for dark mode. Defaults to NULL (auto-set if darkmode = TRUE).}

\item{upper.bound}{Numeric value indicating the upper bound for the y axis. Defaults to -20 to avoid floating y axis ticks. Can be tweaked for aesthetics.}

\item{lower.bound}{Numeric value indicating the lower bound for the y axis. Defaults to "default", which sets it to the maximum depth in the data.}
}
\value{
A ggplot object representing the transect plot.
}
\description{
This function creates transect plots from cGENIE model output .nc files.
}
