% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.transect.R
\name{cGENIE.transect}
\alias{cGENIE.transect}
\title{Create Transect Plots from cGENIE Model Output}
\usage{
cGENIE.transect(
  experiment,
  var,
  slice = 1,
  dims = 3,
  lat.or.lon = "lat",
  year = "default",
  unit.factor = 1,
  min.value = 0,
  max.value = 40,
  intervals = 5,
  continents.outlined,
  scale.label = "[label goes here]",
  model = "biogem",
  palette_name = pals::parula(1000),
  projection = "ESRI:54012"
)
}
\arguments{
\item{experiment}{Directory containing the experiment's netCDF files.}

\item{var}{The variable name to extract from the .nc file.}

\item{slice}{The slice index to extract from the data. Defaults to 1.}

\item{dims}{Number of dimensions in the netCDF file. Defaults to 3.}

\item{lat.or.lon}{Whether to create a transect along latitude or longitude. Defaults to "lat".}

\item{year}{Year to extract data for (default is "default", meaning the last time point).}

\item{unit.factor}{Factor to convert units. Defaults to 1.}

\item{min.value}{Minimum value for the color scale. Defaults to 0.}

\item{max.value}{Maximum value for the color scale. Defaults to 40.}

\item{intervals}{Number of intervals for the color scale. Defaults to 5.}

\item{continents.outlined}{Logical indicating whether to outline continents. Defaults to NULL.}

\item{scale.label}{Label for the color scale. Defaults to "\link{label goes here}".}

\item{model}{The model type; defaults to "biogem".}

\item{palette_name}{Color palette to use for the plot. Defaults to \code{pals::parula(1000)}.}

\item{projection}{Projection to use for the plot. Defaults to 'ESRI:54012'.}
}
\value{
A ggplot object representing the transect plot.
}
\description{
This function creates transect plots from cGENIE model output .nc files.
}
\examples{
\dontrun{
transect_plot <- cGENIE.transect(experiment = "experiment1", var = "temperature")
}
}
