% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.years.R
\name{cGENIE.years}
\alias{cGENIE.years}
\title{Prints the years that are saved in a cGENIE model experiment}
\usage{
cGENIE.years(experiment, model = "biogem")
}
\arguments{
\item{experiment}{Directory containing the experiment's netCDF files.}

\item{model}{The model type; defaults to "biogem".}
}
\value{
A data frame containing the extracted 2D lat-lon data field with corresponding grid information.
}
\description{
Prints the years that are saved in a cGENIE model experiment
}
