% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cGENIE.years.R
\name{cGENIE.years}
\alias{cGENIE.years}
\title{Prints the years that are saved in a cGENIE model experiment}
\usage{
cGENIE.years(experiment, var = "ocn_temp", model = "biogem", dims = 3)
}
\arguments{
\item{experiment}{Directory containing the experiment's netCDF files.}

\item{var}{The variable name to extract from the .nc file.}

\item{model}{The model type; defaults to "biogem".}
}
\value{
A data frame containing the extracted 2D lat-lon data field with corresponding grid information.
}
\description{
Prints the years that are saved in a cGENIE model experiment
}
