% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HADCM3.data.R
\name{HADCM3.data}
\alias{HADCM3.data}
\title{Extract Data from NetCDF Files}
\usage{
HADCM3.data(
  var,
  file,
  experiment,
  depth.level = 1,
  dims = 3,
  time.present = FALSE
)
}
\arguments{
\item{var}{Character. The name of the variable to extract from the NetCDF file.}

\item{file}{Character. The name of the NetCDF file (without the .nc extension).}

\item{experiment}{Character. The path to the directory containing the NetCDF file.}

\item{depth.level}{Numeric. The depth level to extract (default is 1).}

\item{dims}{Numeric. The number of dimensions of the NetCDF file (default is 3).}

\item{time.present}{Logical. Whether the time dimension is present in the NetCDF file (default is FALSE).}
}
\value{
A data frame containing the extracted data, with columns for longitude, latitude, and the specified variable.
}
\description{
This function extracts data from imported NetCDF (.nc) files, specifically designed for use with HADCM3 model outputs.
}
\details{
This function reads in a NetCDF file using the RNetCDF package and extracts the specified variable. It handles both 2D and 3D data, and can adjust longitude values to be within the range of -180 to 180 degrees. The function also filters the data to ensure it falls within realistic geographic bounds.
}
\note{
The function assumes that the NetCDF file follows a specific structure, with variables named "latitude", "longitude", "depth_1", and "t" for time. Adjustments may be needed for different file structures.
}
\examples{
\dontrun{
# Example usage:
df <- HADCM3.data(var = "insitu_T_ym_dpth", file = "o.pgclann", experiment = "~/Valdes2021_HADCM3L/teXPl_444/teXPl_444")
}

}
