% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HADCM3.map.pretty.R
\name{HADCM3.map.pretty}
\alias{HADCM3.map.pretty}
\title{HADCM3.map.pretty}
\usage{
HADCM3.map.pretty(
  file = NULL,
  var = NULL,
  experiment = NULL,
  file_2 = NULL,
  depth.level = 1,
  dims = 3,
  land.opt = "default",
  min.value = NULL,
  max.value = NULL,
  intervals = NULL,
  continents.outlined = NULL,
  scale.label = NULL,
  unit.factor = 1,
  time.present = FALSE,
  scale = "viridis",
  projection = "ESRI:54012",
  calcs = TRUE,
  plot = TRUE,
  for.app = FALSE,
  palette_name_ocean = pals::parula(1000),
  polygons = NULL,
  darkmode = FALSE,
  bg.color = "white",
  fg.color = "black",
  col.labels = NULL
)
}
\arguments{
\item{file}{Name of the netCDF file (without extension). This should generally be ".qrparm".}

\item{var}{Name of the variable to extract from the netCDF file. Default is "insitu_T_ym_dpth".}

\item{experiment}{A string specifying the path to the experiment directory.}

\item{file_2}{Name of the second netCDF file (without extension). should generally be ".qrparm.orog"}

\item{depth.level}{An integer specifying the depth level to be visualized. Default is 1.}

\item{dims}{An integer specifying the dimensions of the netCDF file being read in. Default is 3.}

\item{land.opt}{A string specifying land options. Default is "default".}

\item{min.value}{Minimum value for the color scale.}

\item{max.value}{Maximum value for the color scale.}

\item{intervals}{Intervals for the color scale.}

\item{continents.outlined}{Logical indicating whether continents should be outlined.}

\item{scale.label}{Label for the color scale.}

\item{unit.factor}{A numeric factor to adjust units. Default is 1.}

\item{time.present}{Logical indicating whether time is present in the data. Default is FALSE.}

\item{scale}{A string specifying the color scale to use. Default is "viridis".}

\item{projection}{A string specifying the map projection to use. Default is 'ESRI:54012'.}

\item{calcs}{Logical indicating whether calculations should be performed. Default is TRUE.}

\item{plot}{Logical indicating whether to plot the map. Default is TRUE.}

\item{for.app}{Logical indicating whether the map is for an app. Default is FALSE.}

\item{palette_name_ocean}{Color palette for ocean data visualization. Default is pals::parula(1000).}

\item{polygons}{Spatial polygons data frame if calcs is FALSE.}

\item{darkmode}{Logical indicating whether to use dark mode. Default is FALSE.}

\item{bg.color}{Background color for the plot. Default is "white".}

\item{fg.color}{Foreground color for the plot elements. Default is "black".}

\item{col.labels}{Labels for the color scale. Default is NULL.}
}
\value{
If plot is FALSE - returns a spatial polygons data frame. If plot is TRUE - returns a ggplot object.
}
\description{
This function generates visually appealing maps from imported .nc files such as those from Valdes et al. 2021. It aims to include key information to visualize as much of the earth system as possible at one time.
}
