% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scotese.map.R
\name{scotese.map}
\alias{scotese.map}
\title{Generate a Map from NetCDF File}
\usage{
scotese.map(
  map,
  min.value = -6000,
  max.value = 6000,
  intervals = 2000,
  continents.outlined,
  scale.label,
  scale = "viridis",
  projection = "ESRI:54012",
  darkmode = FALSE,
  bg.col = ifelse(darkmode, "black", "white"),
  fg.col = ifelse(darkmode, "white", "black")
)
}
\arguments{
\item{map}{Character. Path to the NetCDF file.}

\item{min.value}{Numeric. Minimum value for the color scale. Default is -6000.}

\item{max.value}{Numeric. Maximum value for the color scale. Default is 6000.}

\item{intervals}{Numeric. Intervals for the color scale. Default is 2000.}

\item{continents.outlined}{Logical. Whether to outline continents. (Parameter not used in the function body)}

\item{scale.label}{Character. Label for the color scale.}

\item{scale}{Character. Color scale to use. Default is "viridis".}

\item{projection}{Character. Projection type for the map. Default is 'ESRI:54012'.}

\item{darkmode}{Logical. Whether to enable dark mode. Default is FALSE.}

\item{bg.col}{Character. Background color for the map. Default is "black" when darkmode is TRUE.}

\item{fg.col}{Character. Foreground color for the map. Default is "white" when darkmode is TRUE.}
}
\value{
A ggplot object representing the map.
}
\description{
This function creates a map visualization from a given NetCDF file. The map can be customized with various parameters including value ranges, intervals, and projection types.
}
